@@ -14,19 +14,19 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.0.9",
-    "@angular/cdk": "17.0.5",
-    "@angular/common": "^17.0.9",
-    "@angular/compiler": "^17.0.9",
-    "@angular/core": "^17.0.9",
-    "@angular/forms": "^17.0.9",
-    "@angular/material": "17.0.5",
-    "@angular/platform-browser": "^17.0.9",
-    "@angular/platform-browser-dynamic": "^17.0.9",
-    "@angular/platform-server": "^17.0.9",
-    "@angular/router": "^17.0.9",
-    "@angular/service-worker": "^17.0.9",
-    "@angular/ssr": "^17.0.10",
+    "@angular/animations": "^17.1.0",
+    "@angular/cdk": "17.1.0",
+    "@angular/common": "^17.1.0",
+    "@angular/compiler": "^17.1.0",
+    "@angular/core": "^17.1.0",
+    "@angular/forms": "^17.1.0",
+    "@angular/material": "17.1.0",
+    "@angular/platform-browser": "^17.1.0",
+    "@angular/platform-browser-dynamic": "^17.1.0",
+    "@angular/platform-server": "^17.1.0",
+    "@angular/router": "^17.1.0",
+    "@angular/service-worker": "^17.1.0",
+    "@angular/ssr": "^17.1.0",
     "@ngrx/operators": "^17.1.0",
     "@ngrx/signals": "^17.1.0",
     "express": "^4.18.2",
@@ -38,20 +38,20 @@
     "zone.js": "~0.14.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.0.10",
+    "@angular-devkit/build-angular": "^17.1.0",
     "@angular-eslint/builder": "17.2.0",
     "@angular-eslint/eslint-plugin": "17.2.0",
     "@angular-eslint/eslint-plugin-template": "17.2.0",
     "@angular-eslint/schematics": "17.2.0",
     "@angular-eslint/template-parser": "17.2.0",
-    "@angular/cli": "^17.0.10",
-    "@angular/compiler-cli": "^17.0.9",
+    "@angular/cli": "^17.1.0",
+    "@angular/compiler-cli": "^17.1.0",
     "@types/express": "^4.17.17",
     "@types/jasmine": "~5.1.0",
     "@types/node": "^20.11.5",
     "@typescript-eslint/eslint-plugin": "6.19.0",
     "@typescript-eslint/parser": "6.19.0",
-    "autoprefixer": "^10.4.16",
+    "autoprefixer": "^10.4.17",
     "eslint": "^8.56.0",
     "eslint-config-prettier": "^9.1.0",
     "jasmine-core": "~5.1.0",
@@ -64,6 +64,6 @@
     "prettier": "^3.2.4",
     "prettier-plugin-tailwindcss": "^0.5.11",
     "tailwindcss": "^3.4.1",
-    "typescript": "~5.2.2"
+    "typescript": "~5.3.3"
   }
 }
