@@ -30,7 +30,7 @@
     "firebase-tools": "^13.0.3",
     "npm-run-all": "^4.1.5",
     "prettier": "^3.2.4",
-    "typescript": "~5.2.2"
+    "typescript": "~5.3.3"
   },
   "private": true
 }
